add_check_ownership

+++ controllers/Book.js
@@ -30,15 +30,35 @@ exports.getOneBook = (req, res, next) => {
 };
 
 exports.modifyBook = (req, res, next) => {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Livre modifié !'}))
-      .catch(error => res.status(400).json({ error }));
+  Book.findOne({ _id: req.params.id })
+    .then(book => {
+      if (!book) {
+        return res.status(404).json({ error: 'Livre non trouvé.' });
+      }
+      if (book.userId !== req.auth.userId) {
+        return res.status(403).json({ error: 'Requête non autorisée.' });
+      }
+      Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
+        .then(() => res.status(200).json({ message: 'Livre modifié !' }))
+        .catch(error => res.status(400).json({ error }));
+    })
+    .catch(error => res.status(500).json({ error }));
 };
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Livre supprimé !'}))
-      .catch(error => res.status(400).json({ error }));
+  Book.findOne({ _id: req.params.id })
+    .then(book => {
+      if (!book) {
+        return res.status(404).json({ error: 'Livre non trouvé.' });
+      }
+      if (book.userId !== req.auth.userId) {
+        return res.status(403).json({ error: 'Requête non autorisée.' });
+      }
+      Book.deleteOne({ _id: req.params.id })
+        .then(() => res.status(200).json({ message: 'Livre supprimé !' }))
+        .catch(error => res.status(400).json({ error }));
+    })
+    .catch(error => res.status(500).json({ error }));
 };
 
 exports.ratingBook = (req, res, next) => {
